backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
# Shared with frontend (already installed in .venv)
sqlalchemy>=2.0
pymysql>=1.1
python-dotenv>=1.0
passlib[bcrypt]>=1.7
